Recombined modular Python script
